+++ BankApplication/src/bankapplication/SavingsAccount.java
@@ -2,7 +2,7 @@ package bankapplication;
 
 public class SavingsAccount extends Account{
     
-    public SavingsAccount(String name, int balance) {
+    public SavingsAccount(String name, double balance) {
         super(name, balance);
     }
